new commit

+++ PersonRegistrationSystem/DataBase/Entities/PersonalInfo.cs
@@ -8,7 +8,7 @@ namespace PersonRegistrationSystem.DataBase.Entities
         public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int PersonalId { get; set; }
+        public long PersonalId { get; set; }
         public int PhoneNumber { get; set; }
         public string Email { get; set; }
         public byte[] ProfilePhoto { get; set; }
