+++ PersonRegistrationSystem/Dtos/ResponceDtos/PersonalInfoResponceDto.cs
@@ -7,7 +7,7 @@ namespace PersonRegistrationSystem.Dtos.ResponceDtos
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int PersonalId { get; set; }
+        public long PersonalId { get; set; }
         public int PhoneNumber { get; set; }
         public string Email { get; set; }       
         //public byte[] ProfilePhoto { get; set; }
